Updated installer with latest versions.

+++ jupyterhub_traefik_proxy/install.py
@@ -10,24 +10,30 @@ import hashlib
 import warnings
 
 checksums_traefik = {
-    "https://github.com/containous/traefik/releases/download/v1.7.18/traefik_linux-amd64": "3c2d153d80890b6fc8875af9f8ced32c4d684e1eb5a46d9815337cb343dfd92e",
-    "https://github.com/containous/traefik/releases/download/v1.7.18/traefik_darwin-amd64": "84e07a184c31b7fb86417ba3a237ad334a26bcb1ed53fd56f0774afaa34074d9",
-    "https://github.com/containous/traefik/releases/download/v1.7.5/traefik_linux-amd64": "4417a9d83753e1ad6bdd64bbbeaeb4b279bcc71542e779b7bcb3b027c6e3356e",
-    "https://github.com/containous/traefik/releases/download/v1.7.5/traefik_darwin-amd64": "379d4af242743a3fe44b44a1ee6df68ea8332578d85de35f264e062c19fd20a0",
-    "https://github.com/containous/traefik/releases/download/v1.7.0/traefik_linux-amd64": "b84cb03e8a175b8b7d1a30246d19705f607c6ae5ee89f2dca7a1adccab919135",
-    "https://github.com/containous/traefik/releases/download/v1.7.0/traefik_darwin-amd64": "3000cb9f8ed567e9bc567cce33107f6877f2017c69fae8ac235b51a7a94229bf",
+    "https://github.com/containous/traefik/releases/download/v2.2.0/traefik_v2.2.0_linux_amd64.tar.gz":
+        "eddea0507ad715c723662e7c10fdab554eb64379748278cd2d09403063e3e32f",
+    "https://github.com/containous/traefik/releases/download/v2.2.0/traefik_v2.2.0_darwin_amd64.tar.gz":
+        "8bfa2393b265ef01aca12be94d67080961299968bd602f3708480eed273b95e0",
+    "https://github.com/containous/traefik/releases/download/v2.2.0/traefik_v2.2.0_windows_amd64.zip":
+        "9a794e395b7eba8d44118c4a1fb358fbf14abff3f5f5d264f46b1d6c243b9a5e",
 }
 
 checksums_etcd = {
-    "https://github.com/etcd-io/etcd/releases/download/v3.3.10/etcd-v3.3.10-linux-amd64.tar.gz": "1620a59150ec0a0124a65540e23891243feb2d9a628092fb1edcc23974724a45",
-    "https://github.com/etcd-io/etcd/releases/download/v3.3.10/etcd-v3.3.10-darwin-amd64.tar.gz": "fac4091c7ba6f032830fad7809a115909d0f0cae5cbf5b34044540def743577b",
-    "https://github.com/etcd-io/etcd/releases/download/v3.2.25/etcd-v3.3.10-linux-amd64.tar.gz": "8a509ffb1443088d501f19e339a0d9c0058ce20599752c3baef83c1c68790ef7",
-    "https://github.com/etcd-io/etcd/releases/download/v3.2.25/etcd-v3.3.10-darwin-amd64.tar.gz": "9950684a01d7431bc12c3dba014f222d55a862c6f8af64c09c42d7a59ed6790d",
+    "https://github.com/etcd-io/etcd/releases/download/v3.4.7/etcd-v3.4.7-linux-amd64.tar.gz":
+        "4ad86e663b63feb4855e1f3a647e719d6d79cf6306410c52b7f280fa56f8eb6b",
+    "https://github.com/etcd-io/etcd/releases/download/v3.4.7/etcd-v3.4.7-darwin-amd64.zip":
+        "ffe3237fcb70b7ce91c16518c2f62f3fa9ff74ddc10f7b6ca83a3b5b29ade19a",
+    "https://github.com/etcd-io/etcd/releases/download/v3.4.7/etcd-v3.4.7-windows-amd64.zip":
+        "3863ea59bcb407113524b51406810e33d58daff11ca10d1192f289185ae94ffe",
 }
 
 checksums_consul = {
-    "https://releases.hashicorp.com/consul/1.5.0/consul_1.5.0_linux_amd64.zip": "1399064050019db05d3378f757e058ec4426a917dd2d240336b51532065880b6",
-    "https://releases.hashicorp.com/consul/1.5.0/consul_1.5.0_darwin_amd64.zip": "b4033ea6871fe6136ee5d940c834be2248463c3ec248dc22370e6d5360931325",
+    "https://releases.hashicorp.com/consul/1.7.2/consul_1.7.2_linux_amd64.zip":
+        "5ab689cad175c08a226a5c41d16392bc7dd30ceaaf90788411542a756773e698",
+    "https://releases.hashicorp.com/consul/1.7.2/consul_1.7.2_darwin_amd64.zip":
+        "c474f00b022cae38acae2d19b2a707a4fcb08dfdd22875efeefdf052ce19c90b",
+    "https://releases.hashicorp.com/consul/1.7.2/consul_1.7.2_windows_amd64.zip":
+        "e9b9355f77f80b2c0940888cb0d27c44a5879c31e379ef21ffcfd36c91d202c1",
 }
 
 
@@ -41,54 +47,75 @@ def checksum_file(path):
 
 
 def install_traefik(prefix, plat, traefik_version):
-    traefik_bin = os.path.join(prefix, "traefik")
+    plat = plat.replace("-", "_")
+    if "windows" in plat:
+        traefik_archive_extension = "zip"
+        traefik_bin = os.path.join(prefix, "traefik.exe")
+    else:
+        traefik_archive_extension = "tar.gz"
+        traefik_bin = os.path.join(prefix, "traefik")
+
+    traefik_archive = "traefik_v" + traefik_version + "_" + plat + "." + traefik_archive_extension
+    traefik_archive_path = os.path.join(prefix, traefik_archive)
+
+    traefik_archive = "traefik_v" + traefik_version + "_" + plat + "." + traefik_archive_extension
+    traefik_archive_path = os.path.join(prefix, traefik_archive)
+
+    traefik_archive = "traefik_v" + traefik_version + "_" + plat + "." + traefik_archive_extension
+    traefik_archive_path = os.path.join(prefix, traefik_archive)
 
     traefik_url = (
         "https://github.com/containous/traefik/releases"
-        f"/download/v{traefik_version}/traefik_{plat}"
+        f"/download/v{traefik_version}/{traefik_archive}"
     )
 
-    if os.path.exists(traefik_bin):
+    if os.path.exists(traefik_bin) and os.path.exists(traefik_archive_path):
         print(f"Traefik already exists")
         if traefik_url not in checksums_traefik:
             warnings.warn(
-                f"Couldn't verify checksum for traefik-v{traefik_version}-{plat}",
+                f"Traefik {traefik_version} not supported !",
                 stacklevel=2,
             )
             os.chmod(traefik_bin, 0o755)
             print("--- Done ---")
             return
         else:
-            checksum = checksum_file(traefik_bin)
-            if checksum == checksums_traefik[traefik_url]:
+            if checksum_file(traefik_archive_path) == checksums_traefik[traefik_url]:
                 os.chmod(traefik_bin, 0o755)
                 print("--- Done ---")
                 return
             else:
-                print(f"checksum mismatch on {traefik_bin}")
+                print(f"checksum mismatch on {traefik_archive_path}")
+                os.remove(traefik_archive_path)
                 os.remove(traefik_bin)
 
-    print(f"Downloading traefik {traefik_version}...")
-    urlretrieve(traefik_url, traefik_bin)
-
     if traefik_url in checksums_traefik:
-        checksum = checksum_file(traefik_bin)
-        if checksum != checksums_traefik[traefik_url]:
+        print(f"Downloading traefik {traefik_version}...")
+        urlretrieve(traefik_url, traefik_archive_path)
+
+        if checksum_file(traefik_archive_path) != checksums_traefik[traefik_url]:
             raise IOError("Checksum failed")
+
+        print("Extracting the archive...")
+        if traefik_archive_extension == "tar.gz":
+            with tarfile.open(traefik_archive_path, "r") as tar_ref:
+                tar_ref.extract("traefik", prefix)
+        else:
+            with zipfile.ZipFile(traefik_archive_path, "r") as zip_ref:
+                zip_ref.extract("traefik.exe", prefix)
+
+        os.chmod(traefik_bin, 0o755)
     else:
         warnings.warn(
-            f"Couldn't verify checksum for traefik-v{traefik_version}-{plat}",
+            f"Traefik {traefik_version} not supported !",
             stacklevel=2,
         )
-
-    os.chmod(traefik_bin, 0o755)
 
     print("--- Done ---")
 
 
 def install_etcd(prefix, plat, etcd_version):
     etcd_downloaded_dir_name = f"etcd-v{etcd_version}-{plat}"
-    etcd_archive_extension = ".tar.gz"
     if "linux" in plat:
         etcd_archive_extension = "tar.gz"
     else:
@@ -102,7 +129,7 @@ def install_etcd(prefix, plat, etcd_version):
     etcdctl_bin = os.path.join(prefix, "etcdctl")
 
     etcd_url = (
-        "https://github.com/etcd-io/etcd/releases/"
+        "https://github.com/etcd-io/etcd/releases"
         f"/download/v{etcd_version}/etcd-v{etcd_version}-{plat}.{etcd_archive_extension}"
     )
 
@@ -110,7 +137,7 @@ def install_etcd(prefix, plat, etcd_version):
         print(f"Etcd and etcdctl already exist")
         if etcd_url not in checksums_etcd:
             warnings.warn(
-                f"Couldn't verify checksum for etcd-v{etcd_version}-{plat}",
+                f"Etcd {etcd_version} not supported !",
                 stacklevel=2,
             )
             os.chmod(etcd_bin, 0o755)
@@ -118,53 +145,52 @@ def install_etcd(prefix, plat, etcd_version):
             print("--- Done ---")
             return
         else:
-            checksum_etcd_archive = checksum_file(etcd_downloaded_archive)
-            if checksum_etcd_archive == checksums_etcd[etcd_url]:
+            if checksum_file(etcd_downloaded_archive) == checksums_etcd[etcd_url]:
                 os.chmod(etcd_bin, 0o755)
                 os.chmod(etcdctl_bin, 0o755)
                 print("--- Done ---")
                 return
             else:
-                print(f"checksum mismatch on etcd")
+                print(f"checksum mismatch on {etcd_downloaded_archive}")
                 os.remove(etcd_bin)
                 os.remove(etcdctl_bin)
                 os.remove(etcd_downloaded_archive)
 
-    if not os.path.exists(etcd_downloaded_archive):
-        print(f"Downloading {etcd_downloaded_dir_name} archive...")
-        urlretrieve(etcd_url, etcd_downloaded_archive)
-    else:
-        print(f"Archive {etcd_downloaded_dir_name} already exists")
-
-    if etcd_archive_extension == "zip":
-        with zipfile.ZipFile(etcd_downloaded_archive, "r") as zip_ref:
-            zip_ref.extract(etcd_downloaded_dir_name + "/etcd", etcd_binaries)
-            zip_ref.extract(etcd_downloaded_dir_name + "/etcdctl", etcd_binaries)
-    else:
-        with (tarfile.open(etcd_downloaded_archive, "r")) as tar_ref:
-            print("Extracting the archive...")
-            tar_ref.extract(etcd_downloaded_dir_name + "/etcd", etcd_binaries)
-            tar_ref.extract(etcd_downloaded_dir_name + "/etcdctl", etcd_binaries)
-
-    shutil.copy(os.path.join(etcd_binaries, etcd_downloaded_dir_name, "etcd"), etcd_bin)
-    shutil.copy(
-        os.path.join(etcd_binaries, etcd_downloaded_dir_name, "etcdctl"), etcdctl_bin
-    )
-
     if etcd_url in checksums_etcd:
-        checksum_etcd_archive = checksum_file(etcd_downloaded_archive)
-        if checksum_etcd_archive != checksums_etcd[etcd_url]:
+        if not os.path.exists(etcd_downloaded_archive):
+            print(f"Downloading {etcd_downloaded_dir_name} archive...")
+            urlretrieve(etcd_url, etcd_downloaded_archive)
+        else:
+            print(f"Archive {etcd_downloaded_dir_name} already exists")
+
+        if checksum_file(etcd_downloaded_archive) != checksums_etcd[etcd_url]:
             raise IOError("Checksum failed")
+
+        print("Extracting the archive...")
+
+        if etcd_archive_extension == "zip":
+            with zipfile.ZipFile(etcd_downloaded_archive, "r") as zip_ref:
+                zip_ref.extract(etcd_downloaded_dir_name + "/etcd", etcd_binaries)
+                zip_ref.extract(etcd_downloaded_dir_name + "/etcdctl", etcd_binaries)
+        else:
+            with (tarfile.open(etcd_downloaded_archive, "r")) as tar_ref:
+                tar_ref.extract(etcd_downloaded_dir_name + "/etcd", etcd_binaries)
+                tar_ref.extract(etcd_downloaded_dir_name + "/etcdctl", etcd_binaries)
+
+        shutil.copy(os.path.join(etcd_binaries, etcd_downloaded_dir_name, "etcd"), etcd_bin)
+        shutil.copy(
+            os.path.join(etcd_binaries, etcd_downloaded_dir_name, "etcdctl"), etcdctl_bin)
+
+        os.chmod(etcd_bin, 0o755)
+        os.chmod(etcdctl_bin, 0o755)
+
+        # Cleanup
+        shutil.rmtree(etcd_binaries)
     else:
         warnings.warn(
-            f"Couldn't verify checksum for etcd-v{etcd_version}-{plat}", stacklevel=2
+            f"Etcd {etcd_version} not supported !",
+            stacklevel=2
         )
-
-    os.chmod(etcd_bin, 0o755)
-    os.chmod(etcdctl_bin, 0o755)
-
-    # Cleanup
-    shutil.rmtree(etcd_binaries)
 
     print("--- Done ---")
 
@@ -172,14 +198,12 @@ def install_etcd(prefix, plat, etcd_version):
 def install_consul(prefix, plat, consul_version):
     plat = plat.replace("-", "_")
     consul_downloaded_dir_name = f"consul_v{consul_version}_{plat}"
-    consul_archive_extension = ".tar.gz"
     consul_archive_extension = "zip"
 
     consul_downloaded_archive = os.path.join(
         prefix, consul_downloaded_dir_name + "." + consul_archive_extension
     )
     consul_binaries = os.path.join(prefix, "consul_binaries")
-
     consul_bin = os.path.join(prefix, "consul")
 
     consul_url = (
@@ -191,48 +215,44 @@ def install_consul(prefix, plat, consul_version):
         print(f"Consul already exists")
         if consul_url not in checksums_consul:
             warnings.warn(
-                f"Couldn't verify checksum for consul_v{consul_version}_{plat}",
+                f"Consul {consul_version} not supported !",
                 stacklevel=2,
             )
             os.chmod(consul_bin, 0o755)
             print("--- Done ---")
             return
         else:
-            checksum_consul_archive = checksum_file(consul_downloaded_archive)
-            if checksum_consul_archive == checksums_consul[consul_url]:
+            if checksum_file(consul_downloaded_archive) == checksums_consul[consul_url]:
                 os.chmod(consul_bin, 0o755)
                 print("--- Done ---")
                 return
             else:
-                print(f"checksum mismatch on consul")
+                print(f"checksum mismatch on {consul_downloaded_archive}")
                 os.remove(consul_bin)
                 os.remove(consul_downloaded_archive)
 
-    if not os.path.exists(consul_downloaded_archive):
-        print(f"Downloading {consul_downloaded_dir_name} archive...")
-        urlretrieve(consul_url, consul_downloaded_archive)
-    else:
-        print(f"Archive {consul_downloaded_dir_name} already exists")
-
-    with zipfile.ZipFile(consul_downloaded_archive, "r") as zip_ref:
-        zip_ref.extract("consul", consul_binaries)
-
-    shutil.copy(os.path.join(consul_binaries, "consul"), consul_bin)
-
     if consul_url in checksums_consul:
-        checksum_consul_archive = checksum_file(consul_downloaded_archive)
-        if checksum_consul_archive != checksums_consul[consul_url]:
+        if not os.path.exists(consul_downloaded_archive):
+            print(f"Downloading {consul_downloaded_dir_name} archive...")
+            urlretrieve(consul_url, consul_downloaded_archive)
+        else:
+            print(f"Archive {consul_downloaded_dir_name} already exists")
+
+        if checksum_file(consul_downloaded_archive) != checksums_consul[consul_url]:
             raise IOError("Checksum failed")
+
+        with zipfile.ZipFile(consul_downloaded_archive, "r") as zip_ref:
+            zip_ref.extract("consul", consul_binaries)
+
+        shutil.copy(os.path.join(consul_binaries, "consul"), consul_bin)
+        os.chmod(consul_bin, 0o755)
+        # Cleanup
+        shutil.rmtree(consul_binaries)
     else:
         warnings.warn(
-            f"Couldn't verify checksum for consul_v{consul_version}_{plat}",
+            f"Consul {consul_version} not supported !",
             stacklevel=2,
         )
-
-    os.chmod(consul_bin, 0o755)
-
-    # Cleanup
-    shutil.rmtree(consul_binaries)
 
     print("--- Done ---")
 
@@ -245,20 +265,17 @@ def main():
             """\
             Checksums available for:
             - traefik:
-                - v1.7.18-linux-amd64
-                - v1.7.18-darwin-amd64
-                - v1.7.5-linux-amd64
-                - v1.7.5-darwin-amd64
-                - v1.7.0-linux-amd64
-                - v1.7.0-darwin-amd64
+                - v2.2.0-linux-amd64
+                - v2.2.0-darwin-amd64
+                - v2.2.0-windows-amd64
             - etcd:
-                - v3.3.10-linux-amd64
-                - v3.3.10-darwin-amd64
-                - v3.2.25-linux-amd64
-                - v3.2.25-darwin-amd64
+                - v3.4.7-linux-amd64
+                - v3.4.7-darwin-amd64
+                - v3.4.7-windows-amd64
             - consul:
-                - v1.5.0_linux_amd64
-                - v1.5.0_darwin_amd64
+                - v1.7.2_linux_amd64
+                - v1.7.2_darwin_amd64
+                - v1.7.2_windows_amd64
             """
         ),
         formatter_class=argparse.RawTextHelpFormatter,
@@ -307,7 +324,7 @@ def main():
     parser.add_argument(
         "--traefik-version",
         dest="traefik_version",
-        default="1.7.18",
+        default="2.2.0",
         help=textwrap.dedent(
             """\
             The version of traefik to download.
@@ -331,7 +348,7 @@ def main():
     parser.add_argument(
         "--etcd-version",
         dest="etcd_version",
-        default="3.3.10",
+        default="3.4.7",
         help=textwrap.dedent(
             """\
             The version of etcd to download.
@@ -355,7 +372,7 @@ def main():
     parser.add_argument(
         "--consul-version",
         dest="consul_version",
-        default="1.5.0",
+        default="1.7.2",
         help=textwrap.dedent(
             """\
             The version of consul to download.
